PS4 impl C++ -> C (as we don't need GNM anymore)

+++ doomgeneric/doomtype.h
@@ -73,7 +73,7 @@ typedef enum
 {
 	false = 0,
 	true = 1,
-	undef = 0xFFFFFFFF
+	undef = (uint32_t)0xFFFFFFFF
 } boolean;
 
 #endif
